backend/pyproject.toml
[project]
name = "pptxagent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "aiosqlite>=0.21.0",
    "asyncddgs>=0.1.0a1",
    "bs4>=0.0.2",
    "deepagents>=0.2.8",
    "duckduckgo-search>=8.1.1",
    "htbuilder>=0.9.0",
    "httptools>=0.7.1",
    "langchain-experimental>=0.4.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.0",
    "langgraph-checkpoint-redis>=0.2.1",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "markdownify>=1.2.2",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
    "wsproto>=1.3.2",
]
